# pms_api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Message:
    properties:
      message:
        type: string
    type: object
  model.Project:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  model.Task:
    properties:
      author_id:
        type: string
      created_at:
        type: string
      deadline:
        type: string
      description:
        type: string
      executor_id:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      status:
        type: string
      tester_id:
        type: string
    type: object
  model.TaskInserted:
    properties:
      author_id:
        type: string
      deadline:
        type: string
      description:
        type: string
      executor_id:
        type: string
      name:
        type: string
      project_id:
        type: string
      status:
        type: string
      tester_id:
        type: string
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      position:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  model.UserInserted:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      password:
        type: string
      position:
        type: string
      username:
        type: string
    type: object
  model.UserShort:
    properties:
      full_name:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  model.UsersPaged:
    properties:
      items:
        items:
          $ref: '#/definitions/model.UserShort'
        type: array
      page_index:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: PMS - project management system. It is course work of student team
    for 5 course in MEPhI.
  termsOfService: http://swagger.io/terms/
  title: PMS API
  version: "1.0"
paths:
...
  /task:
    post:
      consumes:
      - application/json
      description: Create task
      parameters:
      - description: Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/model.TaskInserted'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Task'
        "422":
          description: Incorrect request body (bind error)
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Create task
      tags:
      - Task
  /task/{id}:
    delete:
      consumes:
      - application/json
      description: Delete task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/model.Message'
        "422":
          description: Incorrect id
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Delete task
      tags:
      - Task
    get:
      consumes:
      - application/json
      description: Get task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Task'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/model.Message'
        "422":
          description: Incorrect id
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Get task by ID
      tags:
      - Task
    put:
      consumes:
      - application/json
      description: Update task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/model.TaskInserted'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Task'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/model.Message'
        "422":
          description: Incorrect id/Incorrect request body (bind error)
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Update task
      tags:
      - Task
  /users:
    get:
      consumes:
      - application/json
      description: Get users
      parameters:
      - description: Page index
        in: query
        name: pageIndex
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UsersPaged'
      security:
      - Login: []
      summary: Get users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserInserted'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
      security:
      - Login: []
      summary: Create user
      tags:
      - User
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete User
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
      security:
      - Login: []
      summary: Delete user
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      security:
      - Login: []
      summary: Get user by ID
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserInserted'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
      security:
      - Login: []
      summary: Update user
      tags:
      - User
  /users/{id}/projects:
    get:
      consumes:
      - application/json
      description: Get user projects
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Project'
            type: array
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/model.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Message'
      security:
      - Login: []
      summary: Get user projects
      tags:
      - User
securityDefinitions:
  Login:
    description: Type "Bearer TOKEN" to correctly set the API Key
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"